fix(review): address PR #977 tenth review round
Address review comments from @MRashad26, copilot-pull-request-reviewer:

- packages/shared/src/interfaces/committee.interface.ts: move
  CommitteeOrganizationFormValue, AcceptInviteOrganizationDialogData,
  AcceptInviteOrganizationDialogResult, and InvitationAcceptContext from
  component/service/utils files into the shared interfaces package per
  project convention (per @MRashad26)
- packages/shared/src/utils/invitation.utils.ts: remove local
  CommitteeOrganizationFormValue interface, import from shared (per @MRashad26)
- accept-invite-organization-dialog.component.ts: remove local interface
  exports, import from @lfx-one/shared/interfaces; change submitting from
  plain boolean to signal(false) so zoneless change detection propagates
  loading/disabled states correctly (per @MRashad26)
- accept-invite-organization-dialog.component.html: update submitting
  bindings to use signal call syntax submitting() (per @MRashad26)
- invitation-accept-flow.service.ts: remove local InvitationAcceptContext
  interface, import from @lfx-one/shared/interfaces (per @MRashad26)
- committee.controller.ts: allow absent (undefined) request body in
  acceptCommitteeInvite; only reject explicitly malformed bodies; normalize
  req.body to {} inside the org-required branch so missing body falls
  through to the existing name validation rather than failing with a
  misleading body-shape error (per copilot-pull-request-reviewer)

Resolves 5 review threads.
