+++ apps/lfx-one/src/server/controllers/committee.controller.ts
@@ -680,7 +680,9 @@ export class CommitteeController {
         return;
       }
 
-      if (req.body === null || typeof req.body !== 'object' || Array.isArray(req.body)) {
+      // Reject bodies that are explicitly present but not a plain object (e.g. arrays, strings).
+      // An absent body (undefined) is allowed — non-org-required accepts send no body at all.
+      if (req.body !== undefined && (req.body === null || typeof req.body !== 'object' || Array.isArray(req.body))) {
         next(
           ServiceValidationError.forField('body', 'Request body must be a JSON object', {
             operation: 'accept_committee_invite',
@@ -699,7 +701,7 @@ export class CommitteeController {
       const acceptData: AcceptCommitteeInviteRequest = {};
 
       if (committeeRequiresOrganization(committee)) {
-        const body = req.body as AcceptCommitteeInviteRequest;
+        const body = (req.body ?? {}) as AcceptCommitteeInviteRequest;
         const orgName = typeof body.organization?.name === 'string' ? body.organization.name.trim() : '';
         if (!orgName) {
           next(
